+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- React `StorageProvider` is now its own function-component, with the `value`
+  property being replaced with `storage` for the storage instance.
 
 ## [0.0.5] - 2019-06-01
 ### Added